Support custom parameters to the authorize URI
Some OIDC providers allow extra paramters, not convered by the default configuration
we pass through.

An example of this is Azure AD, which adds the `resource` parameter which is used to
scope access of the resultant access token.

This adds the ability to pass through custom attributes, as a catch all, to the authorize URI.