Fix handling of JWT without key ID
When a key ID (`kid`) is not included in the JWT, that means we don't
know anything about which signing key to use. The json-jwt library
expects the `kid` value to be present if a JWK Set is presented.  If a
missing `kid` exception is raised, we now iterate through each key to
find one that works.

Closes https://github.com/m0n9oose/omniauth_openid_connect/issues/64